requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
ecdsa==0.13.3             # via python-jose
future==0.18.2            # via python-jose
idna==2.8                 # via requests
prettytable==0.7.2
pyasn1==0.4.7             # via rsa
pyjwt==1.7.1
python-jose==3.0.1        # via python-keycloak-client
python-keycloak-client==0.2.2
requests==2.22.0
rsa==4.0                  # via python-jose
six==1.12.0               # via python-jose
urllib3==1.25.6
